requirements.txt
requests==2.22.0
seaborn==0.9.0
pystempel==1.1.0
wordcloud==1.5.0
pandas==0.25.1
numpy==1.17.2
Pillow==6.2.1
scikit_learn==0.21.3
